Load balancer supports multiple games per server
If no suitable game topics are found, the load balancer
will create a new topic and request a server to start listening
to it. The server is picked out on random from all known servers.
The LB will then give this topic to a client to use for the game